pyproject.toml
[project]
name = "asr-model-benchmark"
version = "0.1.0"
description = "Transcription benchmark for comparing Google Cloud speech-to-text models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-speech>=2.33.0",
    "google-cloud-storage>=3.4.0",
    "google-genai>=1.39.1",
    "jiwer>=4.0.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-pyyaml>=6.0.0",
]
